Prepare for sharing
