+++ Docker/DockerfileRun
@@ -1,0 +1,37 @@
+FROM ubuntu:16.04
+
+# Install base packages
+RUN apt-get update && apt-get install -y \
+    git \
+    wget \
+    vim
+
+# Install Java
+RUN apt-get update
+RUN apt-get -y install openjdk-8-jdk
+RUN export JAVA_HOME="/usr/lib/jvm/java-8-openjdk-amd64"
+RUN export PATH="$JAVA_HOME/bin:$PATH"
+RUN export _JAVA_OPTIONS="-Xmx16g -Xms12g -Dfile.encoding=UTF-8"
+
+# Install Scala
+WORKDIR /
+RUN apt-get install apt-transport-https -y
+RUN echo "deb https://dl.bintray.com/sbt/debian /" |  tee -a /etc/apt/sources.list.d/sbt.list
+RUN apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv 2EE0EA64E40A89B84B2DF73499E82A75642AC823
+RUN apt-get update && apt-get install sbt=1.2.8 -y
+
+# Clone the ConceptAlignment repo
+WORKDIR /conceptalignment-project
+RUN git clone https://github.com/clulab/conceptalignment.git
+
+# Copy files to the image.
+COPY hnswlib-datamart.idx .
+COPY hnswlib-glove.idx .
+COPY lucene-datamart ./lucene-datamart
+
+# Make sure it works/get a head start.
+WORKDIR /conceptalignment-project/conceptalignment
+RUN sbt webapp/compile
+
+# Allow for web service
+EXPOSE 9000
