+++ README.md
@@ -11,7 +11,7 @@ or [dynamically](http://linking.cs.arizona.edu/api).
 
 ## Notes for the initiated
 
-* It will take any words, preferably those found in glove.  Others don't do much good.
+* It will take any words, but preferably those found in glove.  Others don't do much good.
   
 * It presently works independently of ontology.  One could enter the description or
 examples associated with an ontology node like "livestock feed hay CSB silage corn meal
@@ -22,7 +22,13 @@ bottom-up components.
   
 * It only scrapes the ISI datamarts.  NYU is not presently included.  It also does SuperMaaS models.
 
-## Operation
+## Preparations
+
+### Preparing the index files
+
+This instructions are for use on a development machine.  However, once the docker
+container has started up, these commands can also be run there.  Initial versions of all the
+indexes will be available in the container, so these commands would be used for updating.
 
 * Run the scrapers using the ScraperApp, which is configured for the
   * IsiScraper and
@@ -38,3 +44,41 @@ bottom-up components.
 * For the alignment run either the
   * SingleKnnApp directly or access it via the
   * webapp subproject.
+
+```bash
+$ # Scrape the datamarts.
+$ sbt scraper/runMain org.clulab.alignment.scraper.ScraperApp ../datamarts.tsv
+$ # Run this one just once because it takes a long time and glove shouldn't change.
+$ sbt indexer/runMain org.clulab.alignment.indexer.knn.hnswlib.HnswlibGloveIndexerApp
+$ # Run these each time the datamarts have changed.
+$ sbt indexer/runMain org.clulab.alignment.indexer.knn.hnswlib.HnswlibDatamartIndexerApp ../datamarts.tsv
+$ sbt indexer/runMain org.clulab.alignment.indexer.lucene.LuceneIndexerApp ../datamarts.tsv
+$ # Start the server in development mode.
+$ sbt webapp/run
+```
+
+### Preparing the Docker image
+
+```bash
+$ # Copy output files to docker directory.
+$ cp ../hnswlib-datamart.idx ../hnswlib-glove.idx Docker
+$ cp -r ../lucene-datamart Docker
+$ # Create the docker image.
+$ docker build -f DockerfileRun . -t conceptalignment-image
+$ # Make a tar file for exchange.
+$ docker save conceptalignment-image | gzip > conceptalignment-image-1.tar.gz
+```
+
+### Preparing the Docker container
+
+```bash
+$ # Download the image from gdrive.
+$ # Load it.
+$ docker load < conceptalignment-image-1.tar.gz
+$ # Create the container.
+$ docker run -id -p 9000:9000 --name conceptalignment-container conceptalignment-image
+$ # Run a shell in the container.
+$ docker exec -it conceptalignment-container bash
+$ # Start the webapp.
+$ sbt webapp/run
+```
