+++ indexer/src/main/scala/org/clulab/alignment/indexer/knn/hnswlib/HnswlibDatamartIndexerApp.scala
@@ -1,0 +1,7 @@
+package org.clulab.alignment.indexer.knn.hnswlib
+
+object HnswlibDatamartIndexerApp extends App {
+  val hnswlibIndexer = new HnswlibIndexer()
+  
+  hnswlibIndexer.indexDatamart(args(0))
+}
