+++ indexer/src/main/scala/org/clulab/alignment/indexer/knn/hnswlib/HnswlibGloveIndexerApp.scala
@@ -1,0 +1,7 @@
+package org.clulab.alignment.indexer.knn.hnswlib
+
+object HnswlibGloveIndexerApp extends App {
+  val hnswlibIndexer = new HnswlibIndexer()
+
+  hnswlibIndexer.indexGlove()
+}
