+++ scraper/src/main/resources/application.conf
@@ -10,6 +10,7 @@ NyuScraper {
 
 SuperMaasScraper {
   url = "http://192.168.1.105:8000/api/v1" // IP address of virtual machine
+  createdSince = "" // This doesn't seem to be implemented yet at the server.
 }
 
 Scraper {
