+++ webapp/public/openapi/webapp.yaml
@@ -17,17 +17,23 @@ components:
         type: object
         properties:
           score:
+            description: a value in the range [-1, 1] with higher being a better match for the search
             type: number
             format: float
           datamartId:
+            description: an indication of the source of the data to distinguish clashing datasetIds from different sources.  Together the datamartId, datasetId, and variableId should form a GUID.
             type: string
           datasetId:
+            description: a unique identifer for the dataset of a particular datamart
             type: string
           variableId:
+            description: a unique identifier for a variable of a particular dataset
             type: string
           variableName:
+            description: a more human readable version of the variableId
             type: string
           variableDescription:
+            description: an more thorough explanation of the variable
             type: string
             
 paths:
